wekiwi-ai-search-and-feature-extraction/pyproject.toml
[tool.poetry]
name = "kiwi-ai-services-2"
version = "0.0.2"
description = ""
authors = ["Christopher Hörnle <christopher.hoernle@digital-at-work.de>"]
license = "AGPLv3"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">3.10.4,<3.13"
fastapi = "^0.112.0"
asyncpg = "^0.29.0"
uvicorn = "^0.30.5"
httpx = "^0.26.0"
loguru = "^0.7.2"
pymilvus = "2.4.4"
sentence-transformers = "3.0.1"
pydirectus = {git = "https://github.com/johind/pydirectus.git"}
polars = "^1.1.0"
flagembedding = "^1.2.10"
langchain = "^0.2.6"
lxml = "^5.2.2"
bs4 = "^0.0.2"
torch = "^2.4.0"
ftfy = "^6.2.3"
cachetools = "^5.4.0"
fastapi-cache2 = {extras = ["memcache"], version = "^0.2.2"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
prospector = "^1.10.3"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"

[build-system]
requires = ["poetry-core>=1.7.1"]
build-backend = "poetry.core.masonry.api"